Upload the final version of the App
After downloading this file, the user goes to the folder which contains the file LocalJsonChart.xcworkspace. Double Click it, and in the Xcode left navigation Panel, select the IoT_Monitor. Then on the right-hand side navigation Panel, change the project location by clicking the folder icon and going to the folder of this project and select the project file called IoT_Monitor.xcodeproj. After that, select the workspace button on top building bar which name is LocalJsonChart (if it is too complex, please check the short guide video iot_app_project_guide.gif to setup the project). Finally, choose the device as iPhone 13 pro and click build button.